Refactor #90: Split exercise_objectives.py into per-exercise modules
