+++ src/opensim_models/optimization/objectives/__init__.py
@@ -1,0 +1,1 @@
+"""Per-exercise objective modules."""
